Move transform configuration into vector-core
This commit moves the `TransformConfig` down into vector-core. The ambition I've
got is to work closer to the architecture described in Architecture Revisit and
to allow for me to experiment with transforms that directly produce into streams
etc etc. Motivation here being an observation in #8512 that we spend a good deal
of time making small vecs, which happens in the topology.
