+++ lib/vector-core/Cargo.toml
@@ -6,8 +6,8 @@ edition = "2018"
 publish = false
 
 [dependencies]
-async-trait = { version = "0.1", default-features = false }
 async-graphql = { version = "=2.6.4", default-features = false, optional = true }
+async-trait = { version = "0.1", default-features = false }
 atomig = { version = "0.3.1", features = ["derive", "serde"] }
 buffers = { path = "buffers", default-features = false }
 bytes = { version = "1.0.1", default-features = false, features = ["serde"] }
@@ -17,6 +17,8 @@ derivative = { version = "2.2.0", default-features = false }
 dyn-clone = { version = "1.0.4", default-features = false }
 futures = { version = "0.3.16", default-features = false, features = ["std"] }
 getset = { version = "0.1.1", default-features = false }
+http = { version = "0.2.4", default-features = false }
+hyper-proxy = { version = "0.9.1", default-features = false, features = ["openssl-tls"] }
 indexmap = { version = "1.7.0", default-features = false, features = ["serde"] }
 lazy_static = { version = "1.4.0", default-features = false }
 lookup = { path = "../lookup", features = ["arbitrary"] }
@@ -24,6 +26,7 @@ metrics = { version = "0.17.0", default-features = false, features = ["std"]}
 metrics-tracing-context = { version = "0.8.0", default-features = false }
 metrics-util = { version = "0.10.0", default-features = false, features = ["std"] }
 mlua = { version = "0.6.2", default-features = false, features = ["lua54", "send", "vendored"], optional = true }
+no-proxy = { version  = "0.3.1", default-features = false, features = ["serialize"] }
 once_cell = { version = "1.8", default-features = false }
 pest = { version = "2.1.3", default-features = false }
 pest_derive = { version = "2.1.0", default-features = false }
@@ -42,6 +45,7 @@ tracing = { version = "0.1.26", default-features = false }
 tracing-core = { version = "0.1.17", default-features = false }
 tracing-log = { version = "0.1.2", default-features = false }
 tracing-subscriber = { version = "0.2.19", default-features = false }
+typetag = { version = "0.1.7", default-features = false }
 twox-hash = { version = "1.6.0", default-features = false }
 vrl-core = { package = "vrl", path = "../vrl/core", optional = true }
