+++ lib/vector-core/src/config/mod.rs
@@ -1,3 +1,6 @@
+mod global_options;
 mod log_schema;
+pub mod proxy;
 
+pub use global_options::GlobalOptions;
 pub use log_schema::{init_log_schema, log_schema, LogSchema};
