+++ lib/vector-core/src/serde.rs
@@ -1,0 +1,6 @@
+/// Answers "Is it possible to skip serializing this value, because it's the
+/// default?"
+#[inline]
+pub fn skip_serializing_if_default<E: Default + PartialEq>(e: &E) -> bool {
+    e == &E::default()
+}
