+++ src/config/mod.rs
@@ -4,20 +4,19 @@ use crate::{
     event::Metric,
     shutdown::ShutdownSignal,
     sinks::{self, util::UriSerde},
-    sources, transforms, Pipeline,
+    sources, Pipeline,
 };
 use async_trait::async_trait;
 use component::ComponentDescription;
 use indexmap::IndexMap; // IndexMap preserves insertion order, allowing us to output errors in the same order they are present in the file
 use serde::{Deserialize, Serialize};
-use shared::TimeZone;
-use snafu::{ResultExt, Snafu};
 use std::collections::{HashMap, HashSet};
 use std::fmt::{self, Display, Formatter};
-use std::fs::DirBuilder;
 use std::hash::Hash;
 use std::net::SocketAddr;
 use std::path::PathBuf;
+pub use vector_core::config::GlobalOptions;
+pub use vector_core::transform::{DataType, ExpandType, TransformConfig};
 
 pub mod api;
 mod builder;
@@ -27,7 +26,6 @@ mod diff;
 pub mod format;
 mod loading;
 pub mod provider;
-pub mod proxy;
 mod unit_test;
 mod validation;
 mod vars;
@@ -40,9 +38,9 @@ pub use loading::{
     load, load_builder_from_paths, load_from_paths, load_from_paths_with_provider, load_from_str,
     merge_path_lists, process_paths, CONFIG_PATHS,
 };
-pub use proxy::ProxyConfig;
 pub use unit_test::build_unit_tests_main as build_unit_tests;
 pub use validation::warnings;
+pub use vector_core::config::proxy::ProxyConfig;
 pub use vector_core::config::{log_schema, LogSchema};
 
 /// Loads Log Schema from configurations and sets global schema.
@@ -87,89 +85,6 @@ pub struct Config {
     expansions: IndexMap<String, Vec<String>>,
 }
 
-#[derive(Clone, Debug, Default, Deserialize, Serialize, PartialEq)]
-#[serde(default)]
-pub struct GlobalOptions {
-    #[serde(default = "default_data_dir")]
-    pub data_dir: Option<PathBuf>,
-    #[serde(skip_serializing_if = "crate::serde::skip_serializing_if_default")]
-    pub log_schema: LogSchema,
-    #[serde(skip_serializing_if = "crate::serde::skip_serializing_if_default")]
-    pub timezone: TimeZone,
-    #[serde(skip_serializing_if = "crate::serde::skip_serializing_if_default")]
-    pub proxy: ProxyConfig,
-}
-
-pub fn default_data_dir() -> Option<PathBuf> {
-    Some(PathBuf::from("/var/lib/vector/"))
-}
-
-#[derive(Debug, Snafu)]
-pub enum DataDirError {
-    #[snafu(display("data_dir option required, but not given here or globally"))]
-    MissingDataDir,
-    #[snafu(display("data_dir {:?} does not exist", data_dir))]
-    DoesNotExist { data_dir: PathBuf },
-    #[snafu(display("data_dir {:?} is not writable", data_dir))]
-    NotWritable { data_dir: PathBuf },
-    #[snafu(display(
-        "Could not create subdirectory {:?} inside of data dir {:?}: {}",
-        subdir,
-        data_dir,
-        source
-    ))]
-    CouldNotCreate {
-        subdir: PathBuf,
-        data_dir: PathBuf,
-        source: std::io::Error,
-    },
-}
-
-impl GlobalOptions {
-    /// Resolve the `data_dir` option in either the global or local
-    /// config, and validate that it exists and is writable.
-    pub fn resolve_and_validate_data_dir(
-        &self,
-        local_data_dir: Option<&PathBuf>,
-    ) -> crate::Result<PathBuf> {
-        let data_dir = local_data_dir
-            .or_else(|| self.data_dir.as_ref())
-            .ok_or(DataDirError::MissingDataDir)
-            .map_err(Box::new)?
-            .to_path_buf();
-        if !data_dir.exists() {
-            return Err(DataDirError::DoesNotExist { data_dir }.into());
-        }
-        let readonly = std::fs::metadata(&data_dir)
-            .map(|meta| meta.permissions().readonly())
-            .unwrap_or(true);
-        if readonly {
-            return Err(DataDirError::NotWritable { data_dir }.into());
-        }
-        Ok(data_dir)
-    }
-
-    /// Resolve the `data_dir` option using
-    /// `resolve_and_validate_data_dir` and then ensure a named
-    /// subdirectory exists.
-    pub fn resolve_and_make_data_subdir(
-        &self,
-        local: Option<&PathBuf>,
-        subdir: &str,
-    ) -> crate::Result<PathBuf> {
-        let data_dir = self.resolve_and_validate_data_dir(local)?;
-
-        let mut data_subdir = data_dir.clone();
-        data_subdir.push(subdir);
-
-        DirBuilder::new()
-            .recursive(true)
-            .create(&data_subdir)
-            .with_context(|| CouldNotCreate { subdir, data_dir })?;
-        Ok(data_subdir)
-    }
-}
-
 #[derive(Debug, Clone, Copy, Deserialize, Serialize)]
 #[serde(default)]
 pub struct HealthcheckOptions {
@@ -199,13 +114,6 @@ impl Default for HealthcheckOptions {
     }
 }
 
-#[derive(Debug, Clone, PartialEq, Copy)]
-pub enum DataType {
-    Any,
-    Log,
-    Metric,
-}
-
 pub trait GenerateConfig {
     fn generate_config() -> toml::Value;
 }
@@ -227,7 +135,7 @@ pub struct SourceOuter {
     pub acknowledgements: bool,
     #[serde(
         default,
-        skip_serializing_if = "crate::serde::skip_serializing_if_default"
+        skip_serializing_if = "vector_core::serde::skip_serializing_if_default"
     )]
     pub proxy: ProxyConfig,
     #[serde(flatten)]
@@ -327,7 +235,7 @@ pub struct SinkOuter {
 
     #[serde(
         default,
-        skip_serializing_if = "crate::serde::skip_serializing_if_default"
+        skip_serializing_if = "vector_core::serde::skip_serializing_if_default"
     )]
     proxy: ProxyConfig,
 
@@ -465,35 +373,6 @@ pub struct TransformOuter {
     #[serde(flatten)]
     pub inner: Box<dyn TransformConfig>,
 }
-
-#[derive(Deserialize, Serialize, Debug)]
-pub enum ExpandType {
-    Parallel,
-    Serial,
-}
-
-#[async_trait]
-#[typetag::serde(tag = "type")]
-pub trait TransformConfig: core::fmt::Debug + Send + Sync + dyn_clone::DynClone {
-    async fn build(&self, globals: &GlobalOptions) -> crate::Result<transforms::Transform>;
-
-    fn input_type(&self) -> DataType;
-
-    fn output_type(&self) -> DataType;
-
-    fn transform_type(&self) -> &'static str;
-
-    /// Allows a transform configuration to expand itself into multiple "child"
-    /// transformations to replace it. This allows a transform to act as a macro
-    /// for various patterns.
-    fn expand(
-        &mut self,
-    ) -> crate::Result<Option<(IndexMap<String, Box<dyn TransformConfig>>, ExpandType)>> {
-        Ok(None)
-    }
-}
-
-dyn_clone::clone_trait_object!(TransformConfig);
 
 pub type TransformDescription = ComponentDescription<Box<dyn TransformConfig>>;
