+++ src/serde.rs
@@ -2,6 +2,7 @@ use indexmap::map::IndexMap;
 use serde::{de, Deserialize, Serialize};
 use std::fmt;
 use std::marker::PhantomData;
+pub use vector_core::serde::skip_serializing_if_default;
 
 pub fn default_true() -> bool {
     true
@@ -14,12 +15,6 @@ pub fn default_false() -> bool {
 pub fn to_string(value: impl serde::Serialize) -> String {
     let value = serde_json::to_value(value).unwrap();
     value.as_str().unwrap().into()
-}
-
-/// Answers "Is it possible to skip serializing this value, because it's the
-/// default?"
-pub(crate) fn skip_serializing_if_default<E: Default + PartialEq>(e: &E) -> bool {
-    e == &E::default()
 }
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
